[HDCameraEditor] Properly instantiate ColorTarget for preview camera
